+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [Unreleased]
 
 ### Added
+- added ModelEngine dataset
 - added Shortcuts functionality
 - added [scope].[type].[name]= shortcut for setvariable
   - scope can be
@@ -22,17 +23,20 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
     - d for Double
     - s for String
   - name is the name of the variable
-  - `MythicScribe.enableShortcuts` to disable
-- `MythicScribe.enableDashesSuggestions` config to determine if dashes should be suggested
+  - `MythicScribe.enableShortcuts` to enable or disable this feature
+- `MythicScribe.enableFileSpecificSuggestions` config to determine if file specific suggestions (metaskill files, mob files) should be shown
+- `MythicScribe.disableAcceptSuggestionOnEnter` config to set the reminder to disable `editor.acceptSuggestionOnEnter` should be shown
 
 ### Fixed
 - Attribute autocompletions no longer stop default autocompletions from showing up when called
 - Mechanic autocompletions do pop up once default autocompletions are called after a `-` symbol
+- Some keys that should have allowed for mechanic completions and hovers (FurnitureSkills and the likes) now do, in fact, allow for mechanic completions and hovers
+- MechanicLine autocompletions (@, ~) popping up inside attributes when pressing space
+
 
 ## [0.0.2]
 
 ### Added
-
 - Autocompletion for mechanic line components (@, ?, ~)
 - Support for inline metaskills
   - By writing `=[` a new completion will be displayed, allowing you to write down the syntax for inline metaskills